Allow custom grid spacing
The code has been refactored a bit so that an integer representing the
grid spacing is passed from the Python interface deeper into the code.
It is used not only to space the x points differently (non-linearly),
but also to rearrange the gaussian integration points.

Currently the previously-existing linear spacing and a new
log-linear-log spacing are available. I'll add more spacings later after
polishing this a bit more.